Key Updates: Argon2 Implimentation and merging User classes. - Updated the gradle file to remove the extra Argon2 dependencies (not needed w/ springboot) - Updated AuthConfig to use Argon2 instead of Bcrypt - Merged the two user classes into and moved role into entities. - Removed redundent dependencies
