+++ backend/utility/build.gradle
@@ -54,9 +54,6 @@ dependencies {
     implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
     runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
     runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
-
-    // **Argon2 Password Hashing**
-    implementation 'de.mkammerer:argon2-jvm:2.11'
 }
 
 tasks.named('test') {
